backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
httpx>=0.28.0,<1.0.0
sse-starlette>=2.2.0,<3.0.0
supabase>=2.12.0,<3.0.0
daytona>=0.143.0,<1.0.0
openhands-ai>=1.3.0,<2.0.0
PyJWT[crypto]>=2.10.0,<3.0.0
slowapi>=0.1.9,<1.0.0
